requirements.txt
absl-py==0.11.0
aiohttp==3.7.3
aiohttp-cors==0.7.0
aioredis==1.3.1
argon2-cffi==20.1.0
async-generator==1.10
async-timeout==3.0.1
attrs==20.3.0
backcall==0.2.0
beautifulsoup4==4.9.3
bleach==3.2.1
blessings==1.7
cachetools==4.2.0
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
click==7.1.2
colorama==0.4.4
colorful==0.5.4
conda==4.9.1
conda-build==3.20.5
conda-package-handling==1.7.0
cryptography==2.9.2
dataclasses==0.6
decorator==4.4.2
defusedxml==0.6.0
dnspython==2.0.0
entrypoints==0.3
filelock==3.0.12
future==0.18.2
glob2==0.7
google==3.0.0
google-api-core==1.24.1
google-auth==1.24.0
google-auth-oauthlib==0.4.2
googleapis-common-protos==1.52.0
gpustat==0.6.0
grpcio==1.34.0
hiredis==1.1.0
idna==2.9
ipykernel==5.4.2
ipython==7.18.1
ipython-genutils==0.2.0
jedi==0.17.2
Jinja2==2.11.2
json5==0.9.5
jsonschema==3.2.0
jupyter-client==6.1.7
jupyter-core==4.7.0
jupyterlab==2.2.9
jupyterlab-pygments==0.1.2
jupyterlab-server==1.2.0
libarchive-c==2.9
Markdown==3.3.3
MarkupSafe==1.1.1
mistune==0.8.4
mkl-fft==1.2.0
mkl-random==1.1.1
mkl-service==2.3.0
msgpack==1.0.2
multidict==5.1.0
nbclient==0.5.1
nbconvert==6.0.7
nbformat==5.0.8
nest-asyncio==1.4.3
notebook==6.1.5
numpy==1.19.2
nvidia-ml-py3==7.352.0
oauthlib==3.1.0
olefile==0.46
opencensus==0.7.11
opencensus-context==0.1.2
packaging==20.8
pandas==1.1.5
pandocfilters==1.4.3
parso==0.7.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.0.0
pkginfo==1.6.0
prettytable==0.7.2
prometheus-client==0.9.0
prompt-toolkit==3.0.8
protobuf==3.14.0
psutil==5.7.2
ptyprocess==0.6.0
py-spy==0.3.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycosat==0.6.3
pycparser==2.20
Pygments==2.7.1
pyOpenSSL==19.1.0
pyparsing==2.4.7
pyrsistent==0.17.3
PySocks==1.7.1
python-dateutil==2.8.1
python-etcd==0.4.5
pytz==2020.1
PyYAML==5.3.1
pyzmq==20.0.0
ray==1.0.0
redis==3.4.1
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.6
ruamel-yaml==0.15.87
Send2Trash==1.5.0
six==1.14.0
soupsieve==2.0.1
tabulate==0.8.7
tensorboard==2.4.0
tensorboard-plugin-wit==1.7.0
tensorboardX==2.1
terminado==0.9.1
testpath==0.4.4
torch==1.7.0
torchelastic==0.2.1
torchvision==0.8.0
tornado==6.1
tqdm==4.46.0
traitlets==5.0.5
typing-extensions==3.7.4.3
urllib3==1.25.8
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
yarl==1.6.3
